Reorder Preprocessor Definitions #define in PILOT and in the Firmware for better understanding Replace int by correct type Add interactive Console when debbuging Firmware as .exe Console Application ("exit" to quit)

+++ MecapitronicLibrary/Includes/ADC_module.h
@@ -34,7 +34,7 @@
 * Prototypes fonctions
 ****************************************************************************************/
 // Read an ADC channel
-unsigned int ADC_Read(unsigned int);
+uint16 ADC_Read(unsigned int);
 
 
 #endif /* ADC_MODULE_H */
